Update patch set 1
Patch Set 1:

just check that commit reference above. Audio HAL uses 3.0 API. Old devices don't support that API yet. But getPresentationPosition is required for frameworks and need to be implemented. The simple way is add getPresentationPosition to unsupported flag and solve the requirement implementation of it on device repo.

Patch-set: 1